requirements.txt
langchain
langchain-community
openai
streamlit
faiss-cpu
tiktoken
PyPDF2
python-dotenv
streamlit-extras